requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.3
pandas==1.3.4
pandastable==0.12.2.post1
Pillow==8.4.0
requests==2.26.0
selenium==4.1.0
